Unify the server startup
Part of #71
Also related to #70

This PR moves the launching the backend services from the main method to the start_api_server method which removes the code duplication between the `main` function and the integration tests setup.
