+++ crates/app/src/main.rs
@@ -63,48 +63,11 @@ async fn main() -> Result<()> {
     }
 
     let current_backend = load_backend(cm.get_backend_type(), &cm).await?;
-    let readiness_state = current_backend.readiness_state();
-
-    let validator_provider: Arc<dyn ExternalIdentityValidatorProvider + Send + Sync> = current_backend.get();
-
-    let schemas_repository: Arc<SchemaRepository> = current_backend.get();
-    let entities_repository: Arc<PrincipalRepository> = current_backend.get();
-    let identity_repository: Arc<IdentityRepository> = current_backend.get();
-
-    let principal_service = Arc::new(PrincipalService::new(
-        identity_repository.clone(),
-        entities_repository.clone(),
-        schemas_repository.clone(),
-    ));
-
-    let token_provider = Arc::new(TokenService::new(
-        validator_provider.clone(),
-        principal_service.clone(),
-        cm.get_signing_key(),
-        cm.get_key_id(),
-        cm.get_audience(),
-        cm.get_issuer(),
-        cm.get_content_encryption(),
-        MetricsProvider::new(ROOT_METRICS_NAMESPACE, cm.instance_name.clone()),
-    ));
-
-    // The audit_service variable is deprecated, use audit_writer instead.
-    // Will be removed in the next releases.
-    let audit_service: Arc<dyn AuditService> = Arc::new(LogAuditService::new());
-
     let audit_writer: Arc<dyn AuditWriter> = Arc::new(LogAuditService::new());
 
     info!(host:? = &cm.listen_address.ip(); "listening on {}:{}", &cm.listen_address.ip(), &cm.listen_address.port());
 
-    let server = boxer_issuer_http::start_api_server(
-        current_backend,
-        token_provider,
-        audit_service,
-        audit_writer,
-        readiness_state,
-        principal_service,
-        cm,
-    )?;
+    let server = boxer_issuer_http::start_api_server(current_backend, audit_writer, cm, ROOT_METRICS_NAMESPACE)?;
 
     server.await.map_err(anyhow::Error::from)
 }
