+++ crates/boxer-issuer-http/src/lib.rs
@@ -12,13 +12,17 @@ use std::sync::atomic::AtomicBool;
 use utoipa::OpenApi;
 use utoipa_swagger_ui::SwaggerUi;
 
-use crate::services::principal_service::PrincipalServiceTrait;
-use crate::services::token_service::TokenProvider;
+use crate::services::configuration::base::initialization_configuration_manager::InitializationConfigurationManager;
+use crate::services::identity_validator_provider::ExternalIdentityValidatorProvider;
+use crate::services::principal_service::{PrincipalService, PrincipalServiceTrait};
+use crate::services::token_service::{TokenProvider, TokenService};
 use anyhow::Result;
 use boxer_core::http::middleware::audit::audit_recorder::audit_writer::AuditWriter;
 use boxer_core::http::middleware::logging::custom_error_logging;
 use boxer_core::services::audit::AuditService;
+use boxer_core::services::audit::log_audit_service::LogAuditService;
 use boxer_core::services::backends::kubernetes::kubernetes_repository::schema_repository::SchemaRepository;
+use boxer_core::services::observability::open_telemetry::metrics::provider::MetricsProvider;
 use http::controllers::v1;
 use http::health;
 use http::openapi::ApiDoc;
@@ -31,17 +35,35 @@ use services::configuration::models::AppSettings;
 
 pub fn start_api_server(
     current_backend: Arc<dyn IssuerBackend>,
-    token_provider: Arc<dyn TokenProvider>,
-    audit_service: Arc<dyn AuditService>,
     audit_writer: Arc<dyn AuditWriter>,
-    readiness_state: Arc<AtomicBool>,
-    principal_service: Arc<dyn PrincipalServiceTrait>,
     cm: AppSettings,
+    root_metrics_namespace: &'static str,
 ) -> Result<Server, anyhow::Error> {
     let schemas_repository: Arc<SchemaRepository> = current_backend.get();
     let entities_repository: Arc<PrincipalRepository> = current_backend.get();
     let identity_repository: Arc<IdentityRepository> = current_backend.get();
     let identity_provider_repository: Arc<IdentityProviderRepository> = current_backend.get();
+    let validator_provider: Arc<dyn ExternalIdentityValidatorProvider + Send + Sync> = current_backend.get();
+    let readiness_state = current_backend.readiness_state();
+    let principal_service = Arc::new(PrincipalService::new(
+        identity_repository.clone(),
+        entities_repository.clone(),
+        schemas_repository.clone(),
+    ));
+    let token_provider: Arc<dyn TokenProvider> = Arc::new(TokenService::new(
+        validator_provider.clone(),
+        principal_service.clone(),
+        cm.get_signing_key(),
+        cm.get_key_id(),
+        cm.get_audience(),
+        cm.get_issuer(),
+        cm.get_content_encryption(),
+        MetricsProvider::new(root_metrics_namespace, cm.instance_name.clone()),
+    ));
+
+    // The audit_service variable is deprecated, use audit_writer instead.
+    // Will be removed in the next releases.
+    let audit_service: Arc<dyn AuditService> = Arc::new(LogAuditService::new());
 
     info!(host:? = &cm.listen_address.ip(); "listening on {}:{}", &cm.listen_address.ip(), &cm.listen_address.port());
     let server_builder = HttpServer::new(move || {
